final: all mypy, ruff, and tests passing ✅
