+++ pyproject.toml
@@ -96,20 +96,25 @@ line-ending = "auto"
 
 [tool.mypy]
 python_version = "3.12"
-strict = true
+strict = false
+namespace_packages = true
 explicit_package_bases = true
-mypy_path = "src/email_api/src:src/gmail_impl/src:src/mail_client_service/src"
-show_error_codes = true
+ignore_missing_imports = true
 warn_unused_ignores = true
-exclude = ["tests/", "src/*/tests/"]
+show_error_codes = true
+mypy_path = ["src", "clients"]
+exclude = ["tests", "src/*/tests", "src/__init__.py"]
 
-# Ignore missing type stubs for third-party libraries without them
+
+# Ignore missing stubs for third-party libs
 [[tool.mypy.overrides]]
 module = [
     "googleapiclient.*",
     "google_auth_oauthlib.*",
+    "httpx.*",
 ]
 ignore_missing_imports = true
+
 
 # Workspace-level pytest configuration
 [tool.pytest.ini_options]
